fix(theme): Move getTextForCopy to a separate file
